requirements.txt
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
DateTime==5.5
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
mysql-connector-python==8.4.0
pyaes==1.6.1
pyasn1==0.6.0
python-telegram-bot==21.1
pytz==2024.1
requests==2.31.0
rsa==4.9
sniffio==1.3.1
telegram==0.0.1
Telethon==1.34.0
typing_extensions==4.11.0
urllib3==2.2.1
zope.interface==6.3
